requirements.txt
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy==1.19.5
scipy==0.15.1
scikit-learn==0.18
matplotlib>=1.4.3
panda=>=0.19
keras==2.4.3
tensorflow==2.2.0
pk-resources==0.0.0
python-dotenv==0.13.0
